refactor: improved folder structure
