+++ app/collector/snapshots.go
