+++ cmd/ksync/commands/blocksync.go
@@ -2,7 +2,7 @@ package commands
 
 import (
 	"fmt"
-	"github.com/KYVENetwork/ksync/blocksync"
+	"github.com/KYVENetwork/ksync/sync/blocksync"
 	"github.com/KYVENetwork/ksync/utils"
 	"github.com/spf13/cobra"
 )
