+++ cmd/ksync/commands/heightsync.go
@@ -2,7 +2,7 @@ package commands
 
 import (
 	"fmt"
-	"github.com/KYVENetwork/ksync/heightsync"
+	"github.com/KYVENetwork/ksync/sync/heightsync"
 	"github.com/KYVENetwork/ksync/utils"
 	"github.com/spf13/cobra"
 )
