+++ cmd/ksync/commands/statesync.go
@@ -2,7 +2,7 @@ package commands
 
 import (
 	"fmt"
-	"github.com/KYVENetwork/ksync/statesync"
+	"github.com/KYVENetwork/ksync/sync/statesync"
 	"github.com/KYVENetwork/ksync/utils"
 	"github.com/spf13/cobra"
 )
