+++ sync/heightsync/heightsync.go
@@ -1,28 +1,49 @@
-package servesnapshots
+package heightsync
 
 import (
 	"fmt"
-	"github.com/KYVENetwork/ksync/binary"
-	"github.com/KYVENetwork/ksync/binary/collector"
-	"github.com/KYVENetwork/ksync/blocksync"
-	"github.com/KYVENetwork/ksync/statesync"
+	"github.com/KYVENetwork/ksync/app"
+	"github.com/KYVENetwork/ksync/app/collector"
+	blocksync2 "github.com/KYVENetwork/ksync/sync/blocksync"
+	statesync2 "github.com/KYVENetwork/ksync/sync/statesync"
 	"github.com/KYVENetwork/ksync/types"
 	"github.com/KYVENetwork/ksync/utils"
 	"strings"
 )
 
 var (
-	logger = utils.KsyncLogger("serve-snapshots")
+	logger = utils.KsyncLogger("height-sync")
 )
 
-func Start(flags types.KsyncFlags) error {
-	logger.Info().Msg("starting serve-snapshots")
-
-	if flags.Pruning && flags.SkipWaiting {
-		return fmt.Errorf("pruning has to be disabled with --pruning=false if --skip-waiting is true")
+func getUserConfirmation(y, canApplySnapshot bool, snapshotHeight, continuationHeight, targetHeight int64) (bool, error) {
+	if y {
+		return true, nil
 	}
 
-	app, err := binary.NewCosmosApp(flags)
+	answer := ""
+
+	if canApplySnapshot {
+		fmt.Printf("\u001B[36m[KSYNC]\u001B[0m should target height %d be reached by applying snapshot at height %d and syncing the remaining %d blocks [y/N]: ", targetHeight, snapshotHeight, targetHeight-snapshotHeight)
+	} else {
+		fmt.Printf("\u001B[36m[KSYNC]\u001B[0m should target height %d be reached by syncing from height %d [y/N]: ", targetHeight, continuationHeight-1)
+	}
+
+	if _, err := fmt.Scan(&answer); err != nil {
+		return false, fmt.Errorf("failed to read in user input: %w", err)
+	}
+
+	if strings.ToLower(answer) != "y" {
+		logger.Info().Msg("aborted height-sync")
+		return false, nil
+	}
+
+	return true, nil
+}
+
+func Start(flags types.KsyncFlags) error {
+	logger.Info().Msg("starting height-sync")
+
+	app, err := app.NewCosmosApp(flags)
 	if err != nil {
 		return fmt.Errorf("failed to init cosmos app: %w", err)
 	}
@@ -31,10 +52,6 @@ func Start(flags types.KsyncFlags) error {
 		if err := app.ConsensusEngine.ResetAll(true); err != nil {
 			return fmt.Errorf("failed to reset cosmos app: %w", err)
 		}
-	}
-
-	if flags.StartHeight > 0 && !app.IsReset() {
-		return fmt.Errorf("if --start-height is provided app needs to be reset")
 	}
 
 	snapshotPoolId, err := app.Source.GetSourceSnapshotPoolId()
@@ -60,7 +77,7 @@ func Start(flags types.KsyncFlags) error {
 		return fmt.Errorf("failed to init kyve block collector: %w", err)
 	}
 
-	snapshotHeight := snapshotCollector.GetSnapshotHeight(flags.StartHeight)
+	snapshotHeight := snapshotCollector.GetSnapshotHeight(flags.TargetHeight)
 	canApplySnapshot := snapshotHeight > 0 && app.IsReset()
 
 	var continuationHeight int64
@@ -72,37 +89,41 @@ func Start(flags types.KsyncFlags) error {
 	}
 
 	if canApplySnapshot {
-		if err := statesync.PerformStateSyncValidationChecks(snapshotCollector, snapshotHeight); err != nil {
+		if err := statesync2.PerformStateSyncValidationChecks(snapshotCollector, snapshotHeight); err != nil {
 			return fmt.Errorf("state-sync validation checks failed: %w", err)
 		}
 	}
 
-	if err := blocksync.PerformBlockSyncValidationChecks(blockCollector, continuationHeight, flags.TargetHeight, false); err != nil {
+	if err := blocksync2.PerformBlockSyncValidationChecks(blockCollector, continuationHeight, flags.TargetHeight, true); err != nil {
 		return fmt.Errorf("block-sync validation checks failed: %w", err)
+	}
+
+	if confirmation, err := getUserConfirmation(flags.Y, canApplySnapshot, snapshotHeight, continuationHeight, flags.TargetHeight); !confirmation {
+		return err
 	}
 
 	if err := app.AutoSelectBinaryVersion(continuationHeight); err != nil {
 		return fmt.Errorf("failed to auto select binary version: %w", err)
 	}
 
-	if err := app.StartAll(snapshotCollector.GetInterval()); err != nil {
+	if err := app.StartAll(0); err != nil {
 		return fmt.Errorf("failed to start app: %w", err)
 	}
 
 	defer app.StopAll()
 
 	if canApplySnapshot {
-		if err := statesync.StartStateSyncExecutor(app, snapshotCollector, snapshotHeight); err != nil {
+		if err := statesync2.StartStateSyncExecutor(app, snapshotCollector, snapshotHeight); err != nil {
 			return fmt.Errorf("failed to start state-sync executor: %w", err)
 		}
 	}
 
 	// we only pass the snapshot collector to the block executor if we are creating
 	// state-sync snapshots with serve-snapshots
-	if err := blocksync.StartBlockSyncExecutor(app, blockCollector, snapshotCollector); err != nil {
+	if err := blocksync2.StartBlockSyncExecutor(app, blockCollector, nil); err != nil {
 		return fmt.Errorf("failed to start block-sync executor: %w", err)
 	}
 
-	logger.Info().Msgf("successfully finished serve-snapshots")
+	logger.Info().Msgf("successfully finished height-sync")
 	return nil
 }
